fix: move getCharacter function
이미지 반환하는 함수를 로직으로 분리
